Removed 'Transpose Song' instrument check
The 'Transpose Song' function checks if a note is attached to an instrument before attempting to transpose it, meaning that notes on rows with no instrument data (or with the legato ('&&') instrument) would not be transposed. Since I couldn't think of any instances where this check was actually necessary (and I *did* think of 2 instances where the check obstructed the function's purpose), I removed it.